convert unit te4st assembly to multi target framework assembly and delete redundant projects

+++ BionicCode.Net/BionicCode.Utilities.UnitTest.Net/ReflectionExtensionMethodsTests/Resources/Public/Generic/TestClassWithInterfaceAndBaseClass.cs
@@ -2,7 +2,7 @@
 {
   using System;
 
-  public class TestClassWithInterfaceAndBaseClass<T> : TestClassBase<T, int>, ITestClass1<T, UInt128, TestClassWithInterface>
+  public class TestClassWithInterfaceAndBaseClass<T> : TestClassBase<T, int>, ITestClass1<T, int, TestClassWithInterface>
     where T : class
   {
   }
